requirements.txt
groq>=0.9.0
python-dotenv>=1.0.0
requests>=2.31.0
pyaudio>=0.2.14
pydub>=0.25.1
audioop-lts>=0.2.1
praw>=7.7.1
pygame>=2.5.0
rapidfuzz>=3.6.0
